Made SignatureTree Bidirectional
SignatureTree so far has always only considered signature indices where a dimension was appended to existing signature indices. This is inherently asymmetric and got cleared up in this version. Now, also signature indices where a dimension is prepended to existing signature indices are considered. Remark: The class SignatureTreeLinear is not yet fully functional. Do not use it.