Implementation of Elimination Mode
Implemented the elimination mode which only includes new nodes if they improve the performance on the validation set, and discards the weaker half of the leaves. This mode will likely finish training after a finite amount of steps. Removed the SignatureTreeLinear class and made this into a mode of the SignatureTree class.