Implemented first draft of SignatureForest
A SignatureForest is a collection of SignatureTrees, and the relationship between SignatureForests and SignatureTrees is similar to the relationship between Random Forests and Random Trees, although SignatureTrees are of course not decision trees. A SignatureForest may be a useful concept as it avoids the high variance of SignatureTrees. This is only a very rough first implementation. While it works, further functionality needs to be added, and certain improvements can still be made.